Tweaked Mana Star behavior to include any instan...
ce of IManaBlock, not just mana pools. Also allows manastars to check
the block above. This change was made to allow Mana Stars to be used
more easily by addons for non-pool blocks. Specifically, for Botanical
Workshop's Gateway, which is a multiblock structure for crafting and
other things that contains a bunch of mana pools that can be accessed
collectively:
https://github.com/Lazersmoke/Botanical-Workshop/blob/master/src/main/java/lazersmoke/botanicalworkshop/common/block/tile/TileGatewayCore.java