Possible unbindig  fix and a couple optimizations
First the optimizations: Doing the same checks over and over is unhealthy. Its preferred to have a uglier but more efficient code like in writeToPacketNBT.

The possible fix, in the linkCollector you could end up with the linkCollector a null and the needsNew at false. All it was needed was for the cached coords to be invalid and not returning a block. Also, when reading theNBT packet, if the coords are 0,-1,0 just set the cache to null. Its an invalid cache coord, no need to force a call to the world that will fail returning a block.