Update en_US.lang
It started with "ubiquitous", and ended in AAAAAAAAAAAAAA. :)

This is a rough first-pass correction of typos, misspellings, grammar, comma-spliced run-on sentences, awkward or misleading phrasing, and some malapropism and catachresis that wasn't obviously a deliberate stylistic choice. This is by no means a comprehensive revision, and is just the most visible errors I noticed on first reading (at least 236 according to GitHub, or 382 by diffchecker).

Further corrections are certainly possible, but the process might stray too far towards "rewriting", rather than merely "cleaning up". As it stands, I've attempted to preserve the structure and content of the original text as much as possible.

This should be looked over by the original author before merging, for the following reasons:
- To ensure that I haven't "fixed" anything that wasn't deliberately "mis-written" for stylistic effect or to preserve a memetic reference, and
- To pass judgement on a few "errors" that I noticed but did not change, for various reasons, listed below.
- To make sure I haven't "fixed" something wrongly. I'm not a professional writer; I have made, can make, and likely will make mistakes of my own.

Issues Not Corrected (but should be looked at):
- "Soujourner" should be "Sojourner", but this may be an intentional misspelling.
- In the Bellethorne entry, the word "malevelous" - is that supposed to be "Marvelous", "Malevolent", or a deliberate portmanteau?
- "Simpleton" is a noun, describing a "simple" person, with an emphasis on "person". The word is (mis?)used in place of the adjective "Simple". Stylistic affectation, or error?
- "Message from Elven Garde" and "Elven Lore" texts. Several errors, but Elves may have their own grammatical/syntactic quirks, due to who-knows-how-many-thousands of years of isolation and linguistic drift, followed by an Elvish-to-English translation likely done with no native English speakers at hand.