Add oredict support 1.8 stone recipes, closes #1660
the code was copy-pastaed from the addStairsAndSlabs helper function to
hack in oredict support, feel free to just add it to the helper function
if you want though. Looking at other Botania code, this is what comes to
mind though.

see #1703
There is also a very big problem that I marked with a comment that says
"very big problem", where on many many different stone types (end,
metamorph, 1.8) the recipe for stone bricks and for chisled is the same,
so usually you can't craft chisled.

sorry for long commit message