Updating ISSUE_TEMPLATE.MD (#2558)
* Added notice section, formatted with linebreaks

Notice section on top give instructions directing people to the forums or reddit to post ideas as indicating by the contributing.md, because missing the contributing.md file is incredibly easy. I did it, my apologies to Vazkii and Willie. Figured out how to force blood line breaks in the markdown and addeed said line breaks at appropriate locations. Also added a tip at the top so people who post issues know how to do so, because I didn't.

* Commented out the Notice section

Because it works and means that even if someone forgets to delete the notice section, it still won't be seen!