in relation to bug #2879
this commit adds a stack limit to craftingrecipes which have a output >1 (e.g. coal, redstone, glowstone dust duping) to pretend spawning of insane amounts of item
entities in the world.

only applies if a recipe would output more than 1 item
only checks 2 block radius around the mana pool for items
can be disabled if the config value is set to 0 (defaults to max. 64 entities/stacks around the pool)