Fix deadlock with corporea indexes shortly after second world load
Loading a corporea index will place it in a weak set that will be
checked for when a player sends a chat message.
If you leave a world while in range of a corporea index, and chat
quickly after entry (before the garbage collector clears the weak set),
the integrated server will deadlock on getting the chunk from
the stale index TE's world.

Closes williewillus/botania-fabric-issues#33