Fix Entropic lens behaviour not matching TNT
We were checking ourselves and passing the player in as the exploder,
but there is already logic within the Explosion to do that for proper
damage attribution.
The Explosion has logic to skip damage completely for the exploder,
which is usually the PrimedTNT entity in vanilla. We were passing in the player
ourselves which skips damage checks completely.