[Fabric] Fix floating flower models
Uses a custom injection that emulates what we have on Forge, instead of Fabric's loader
API, which is unwieldy. One of the problems is that it dispatches "externally", based
on the requested model's path, instead of "internally", using data in the model itself.

Forge has a better design here.
One workaround I had to use was duplicating the models for items, since "parent" doesn't
work properly. Wastes a bit of memory after baking, but I don't care too much.