Replace a bunch of uses of our Vector3 with vanilla's Vec3
The eventual goal is to be able to drop ours, but there's a few complicated uses
that vanilla code still doesn't support nicely, that we need helpers for.

We want to drop ours to lower maintenance costs and to benefit from sharing
any JIT optimizations with vanilla's very hot copies.