Fix caps being invalidated too early, closes #3992
My assumption was that !entity.isAlive <=> caps are invalidated,
but that isn't true. In the mixin that calls ItemKeepIvy.keepDropsOnDeath,
the caps are actually still available, so I ended up invalidating them early
on accident. This manifests in all LivingDropEvent handlers not being able to
see any caps. The fix is to only do the revive/invalidate dance when
needed, which means the respawn event.

Have I mentioned how stupid this whole idea is of hiding the caps on death??