Migrate to FabricItemGroupBuilder on Fabric
This fixes an incompatibility with Quilt, which does not implement
ItemGroupExtensions.

FabricItemGroupBuilder only accepts a ResourceLocation for the creative
tab name (for some god forsaken reason), so the translation key of the
creative tab has changed from "itemGroup.botania" to "itemGroup.botania.tab.
A separate commit will change the language key in non-en_us language
files.