Fix duplication glitch when mossifying a petal apothecary
We saved and restored the complete NBT state of the old apothecary to the new one. This
ends up duplicating items because a block change causes the contents to drop already, then
we store a duplicate copy of all items into the new apothecary.

The block of code here was originally introduced in
830b87db408c34c30737e7fa580448f9d6187e39 to fix the fluid of the apothecary disappearing
when mossifying, back when the fluid was stored in NBT state. Now the fluid is stored as
part of the blockstate, so this logic is unnecessary and serves solely to cause this bug.

Closes #4057