Prebake all mana blaster and lens models
Before, we were holding onto the ModelBakery and thus keeping alive all the unbaked models and jsons.
With Botania and vanilla alone, this was consuming an extra 16M of memory, but could go to hundreds
of MB on larger packs. Prebaking eliminates most of that memory usage, and leaves us with about
380KB of memory used by the baked models. Closes #4095