Don't exclude botania:livingwood from the pure daisy livingwood_log recipe
In my opinion, we only need to exclude the livingwood log, so that it doesn't repeatedly
convert into itself. Converting livingwood to livingwood logs is fine (the same way it
already converts stripped livingwood/logs back into normal livingwood logs. I view it as
a "purifying process", in a way).