Rework corporea index chat interception
The message is now intercepted clientside instead of serverside, and
we send our own packet instead. This has the benefit of bypassing all
the message signing stuff for chats. It introduces a small window where
you can potentially be in range clientside but not serverside, but that's
fine given that the C-button requests already had that limitation as well.