Hold Corporea networks as sets instead of list
Reading the code, I don't understand why this was ever a list,
since we never provided any guarantees about what nodes service a request first.
Using a set allows for fast membership checks and removes the need for our bespoke
list-set hybrid thing.
For a bit of consistency with before, I use a LinkedHashSet for now,
but that's not guaranteed to stay.