Match reagents in TileAltar
This causes a pretty significant change to how apothecaries work.
Before, apothecaries always rejected any items matching the "seedlike" regex.
Now, such items will enter the apothecary, unless the apothecary 1) has a
complete recipe in itself already and 2) the incoming item matches the recipe's reagent.
The apothecary will always prefer to complete a recipe when it can instead of
adding an item to the water.