Match reagents in PetalApothecaryBlockEntity
This causes a somewhat significant change to how apothecaries work.
Before, apothecaries always rejected any items matching the "seedlike" regex.
Now, such items will enter the apothecary, unless the apothecary 1a) has a
complete recipe in itself already and 1b) the incoming item matches the recipe's reagent.
Or 2) the item is the reagent of the current stored recipe (the latest crafted recipe up
to 20 seconds ago, given the apothecary is currently empty).
The apothecary will always prefer to complete a recipe when it can instead of
adding an item to the water.

Co-authored-by: Vincent Lee <vincent@vincent-lee.net>