Make spectranthemum cost scale with distance
Existing setups will draw the same or less amounts of mana; this patch
simply uncaps the distance limit.

Current balance is such that a full stack can be sent ~40 blocks and a
single item can be sent 2500 for the cost of ~half a diluted pool.
(Subject to change if abused.)