Only kill bursts serverside
One potential source of desyncs has been handling of shouldKill differing between client and server.
This commit makes it so that we only ever issue discard() on bursts on the server. Client effects are handled
by sending EntityEvent.DEATH. This puts the burst inline with all other projectiles in vanilla and Botania.

With this, the fix in 53d38832f0ecb54cc032521f57a0d82efdd98bad may no longer be necessary, this will be
investigated in a followup.