Persist the alreadyCollidedAt array in mana bursts
This thing was never saved, but it's used for state checks in the bore lens. This could lead to
extremely obscure glitches that happen if the chunk is unloaded in the middle of a burst's flight.
ALso reuse setCollidedAt, which handles mutable blockposes properly. There was probably a bug here too.