Update forge
In build 60, Forge patched the FlowerBlock constructor to take a Supplier for the MobEffect.
Unfortunately, it made the non-supplier vanilla constructor crash for all modded effects, because it
attempts to get the registry delegate for that effect, which requires the effect to be registered.
However, modded effects are always registered after blocks on Forge.
This requires a stupid, stupid workaround where we just copy the class per loader, just so we can have a () ->
in front of it on the Forge side. Stupid.