Ensure red-stringed containers always report an accessible inventory (#4308)
Fix for Botania issue #4062, as suggested.

For Forge, the red stringed container reports an empty inventory even
when it's not bound to an inventory.

For Fabric, this was already the case, but the inventory reported that
it absolutely never supports extraction, which is what Corporea sparks
check for the up direction. Since even vanilla Minecraft supports
dynamically adding and removing inventory blocks (shulker boxes) in the
line of sight of the red stringed container, the block's inventory
access options could change at any time and cause e.g. some kind of
pipe-like connection to break. The obvious downside is that a player
could make pipe-like connections or place a Corporea spark on a red
stringed container without the bound inventory actually allowing access
from that side.