Rannuncarpus can now add more candles and sea pickles to matching existing blocks (#4329)
Fix for #4306. Rannuncarpus will do the complete replacement check that
block items do when a player attempts to place them, not just against
replaceable materials. That allows them to place multiple sea pickles or
matching candles on the same block, and should be compatible with any
other block that would allow such placement.