Allow the flower model for floating flowers to use forges model extensions. (#4359)
This changes `ForgeFloatingFlowerModel` such that it uses the
`ChunkRenderTypeSet` from the flower model to construct its block render
types. Also as the floating flower model may add more render types to
the flower model it makes sure, that `getQuads` is only called on the
flower model if the flower model supports the queried render type.

This change allows `SimpleBakedModel` instances with a render type as
well as multi-rendertype models to be used with the floating flower
model loader.