Worldshaper's Sextant sphere mode and finer radius control (#4333)
Implementation for #4312.
- Sphere mode can be toggled by sneak+right-clicking while the sextant
doesn't show a preview.
- Both the sphere *and* circle mode generate a thin shape now,
containing only blocks that are visible from the outside. (The old
version of the circle mode didn't necessarily do that all the time.)
Since there are multiple reasonable ways to represent a voxel sphere or
circle of the same general radius, the sextant now also accepts higher
precision radii, which can cause the shape to vary slightly. Radius
selection is still restricted to discrete values by snapping to a
block-sized grid. (It did that before already, but also rounded down.)
- And finally, this PR restores the particle effects that should be
displayed while selecting a radius and makes them slightly fancier.