Show Manufactory Halo active status in its icon
The original item texture is now its "active" icon, which is still the default state, but has been moved to a texture with the suffix "_active". As a result, 3rd-party resource packs that change the icon will need to be updated.
The base texture (for the inactive state) has been replaced with "programmer art" that looks like a mix of the assembly halo and manufactory halo icons. (The two icons should probably be replaced with ones that include a proper state indicator.)