Upload zh_cn files (#4366)
The main work is to sort the entries, add missing entries and fix some
mistranslations.
In addition, it makes the number of zh_cn lines aligned with en_us,
which is convenient for future Chinese translation. Currently in this
file, the number of zh_cn lines = en_us + 2
The reason for this is that Chinese files have two more lines of opening
comments than English.