Ensure that Astrolabe doesn't do its work "for free"
If blocks are provided by mana-consuming items, calculations for the number of available blocks may be off if the player is short on mana.
As long as preliminary checks suggest that enough blocks should be available, the Astrolabe will now consume its operation mana first and then only place the blocks that can actually be conjured with the remaining mana.