Tag support for Drum of the Gathering (closing #4453)
The drum now uses entity tags to determine which mobs can fill milk buckets and which shearable mobs can NOT be sheared: `botania:drum/milkable` (defaults to cow, mooshroom, and goat) and `botania:drum/no_shearing` (defaults to only mooshroom)
Additionally, if a brown mooshroom was fed a flower, the Drum of the Gathering will fill the first bowl for that mooshroom with the appropriate type of suspicious stew. Any additional bowls will be filled with mushroom stew as usual.
And finally, baby animals are excluded from filling buckets or bowls. (Whether they can be sheared is already determined by the mob's shearing logic, which usually excludes babies as well.)