Prevent Animated Torch from emitting signal before its block entity is initialized
Fixes #4489, where (on Fabric) the torch initially provides power to the south if placed any other direction. The torch now initially does not output a signal until its block entity is initialized, at which point neighbors are updated accordingly.
This also fixes the torch updating too few block when removed (e.g. if it was powering something through a solid block), and unnecessary updates to unaffected blocks when it starts or finishes rotating.