Workaround for floating flowers in multiblock visualization
"Fix" for #4181, by switching floating flowers to block model render shape when displaying them as part of a multiblock visualization. (As if setting the simpleFloaters client option, but only for individual "virtual" blocks.) Potentially also fixes other uses of that particular way to render blocks, such as the block held by an enderman.