Adjust rune recipes for coherence
- water rune: replace fishing rod with lily pad (tools as ingredients feel weird)
- fire rune: replace nether wart with coal or charcoal (i.e. all ingredients obtainable before going to the nether)
- earth rune: replace coal block with mud block (because coal is for fire now; mud is renewable via the Rod of the Lands)
- remove multiple instances of ingredients from season runes and consistently have 3 non-rune ingredients
- spring rune: replace wheat (not a spring crop) with dandelion and carrot
- autumn rune: add beetroot (an autumn crop)
- winter rune: replace cake with sweet berries (a cold crop biome; cake feels too complex and somewhat unrelated)

The change to the fire rune recipe means that the spring and autumn runes, as well as the gluttony, greed, sloth, and pride runes are no longer nether-gated either, unlocking a lot of flowers. The "sin runes" should probably receive nether gating in some other way. Also, this means terrasteel and Alfheim are no longer nether-gated either, which needs to be rectified later.