Force nether-gating for terrasteel
By replacing the manasteel block in the terra plate with a mana quartz block, you spend a less mana on the ingredients, but require nether quartz. This gates the plate itself (and thus the availability of terrasteel and the Alfheim portal) behind a nether visit, but does not require a constant supply of quartz to make terrasteel.