Add REI support for NeoForge version
This is slightly janky, since there is no mojmap "-common" dependency, so we depend on the NeoForge version in Xplat.
There also needs to be a separate subclass of the plugin in the NeoForge module, because the annotation required to register it on NeoForge does not exist oin Fabric, where we can just reference the now-Xplat implementation in fabric.mod.json as before.
While I'm at it, I renamed the Sodium and Embeddium integration packages to live in the `vazkii.botania.neoforge.client.integration` package, where the `BotaniaREINeoforgePlugin` found its home as well.

(closes #4239)